get_file_names

+++ src/constants.f90
@@ -2,6 +2,6 @@ module constants
     use iso_fortran_env, only: real32, real64, real128
     implicit none
     
-    integer, parameter :: pr = real64
+    integer, parameter :: pr = real32
 
 end module constants
